test: share query validation across adapters
