+++ server/index.ts
@@ -55,6 +55,7 @@ import type {
 } from '../src/types/index.js'
 import { buildReconciliationRecord, type ReconciliationRoute } from '../src/lib/reconciliation.js'
 import { appendReconciliationRecord } from './reconciliationStore.js'
+import { validateQuery, MAX_QUERY_LENGTH } from '../src/lib/queryValidation.js'
 
 dotenv.config()
 
@@ -394,28 +395,7 @@ function recordReconciliation(params: {
   }
 }
 
-export const MAX_QUERY_LENGTH = 256
-
-// Validate and sanitize the user-supplied `q` parameter. Returns either the
-// cleaned string or a 400 response body to send back. Centralised so /search
-// and /images share the same rules.
-export function validateQuery(
-  q: unknown,
-): { ok: true; cleanQ: string } | { ok: false; error: string } {
-  if (typeof q !== 'string' || !q.trim()) {
-    return { ok: false, error: 'Missing required parameter: q' }
-  }
-  if (q.length > MAX_QUERY_LENGTH) {
-    return { ok: false, error: `Query too long. Maximum ${MAX_QUERY_LENGTH} characters.` }
-  }
-  // Strip null bytes and ASCII control characters (C0 + DEL) to prevent
-  // log injection and odd Serper behavior.
-  const cleanQ = q.replace(/[\x00-\x1F\x7F]/g, '').trim()
-  if (!cleanQ) {
-    return { ok: false, error: 'Query contains no valid characters.' }
-  }
-  return { ok: true, cleanQ }
-}
+export { validateQuery, MAX_QUERY_LENGTH }
 
 // ─── GET /search ──────────────────────────────────────────────────────────
 app.get('/search', async (req: Request, res: Response) => {
